+++ packages/react/src/modules/core/icons/CheckmarkIcon.tsx
