+++ packages/react/src/modules/core/icons/CloseIcon.tsx
