+++ packages/react/src/modules/core/icons/HashtagIcon.tsx
