+++ packages/react/src/modules/core/icons/InfoIcon.tsx
