+++ packages/react/src/modules/core/icons/LockIcon.tsx
