+++ packages/react/src/modules/core/icons/SearchIcon.tsx
